+++ README.md
@@ -10,7 +10,7 @@
 
 
 ## Installing
-This is a zero-dependency package. It requires Go version 1.22 or above.
+This is a zero-dependency package. It requires Go version 1.21 or above.
 ```shell
 go get github.com/glossd/fetch
 ```
@@ -108,9 +108,9 @@ if resp.Status == 200 {
     fmt.Println("Response headers", resp.Headers)
 }
 ```
-If you don't need the HTTP body you can use `fetch.ResponseEmpty`
+If you don't need the HTTP body you can use `fetch.Empty` or `fetch.Response[fetch.Empty]` to access http attributes
 ```go
-res, err := fetch.Delete[fetch.ResponseEmpty]("https://petstore.swagger.io/v2/pet/10")
+res, err := fetch.Delete[fetch.Response[fetch.Empty]]("https://petstore.swagger.io/v2/pet/10")
 if err != nil {
     panic(err)
 }
@@ -287,13 +287,13 @@ type Config struct {
 ```
 
 ## HTTP Handlers 
-`fetch.ToHandlerFunc` converts `func(in) (out, error)` signature function into `http.HandlerFunc`.
-It unmarshals the HTTP request body into the function argument then marshals the returned value into the HTTP response body.
+`fetch.ToHandlerFunc` converts `func(in) (out, error)` signature function into `http.HandlerFunc`. It does all the json and http handling for you. 
+The HTTP request body unmarshalls into the function argument. The return value is marshaled into the HTTP response body.
 ```go
 type Pet struct {
     Name string
 }
-http.HandleFunc("POST /pets", fetch.ToHandlerFunc(func(in Pet) (*Pet, error) {
+http.HandleFunc("/pets", fetch.ToHandlerFunc(func(in Pet) (*Pet, error) {
     if in.Name == "" {
         return nil, fmt.Errorf("name can't be empty")
     }
@@ -303,21 +303,28 @@ http.ListenAndServe(":8080", nil)
 ```
 If you don't need request or response body, use `fetch.Empty` to fit the function signature.
 ```go
-http.HandleFunc("GET /default-pet", fetch.ToHandlerFunc(func(_ fetch.Empty) (Pet, error) {
+http.HandleFunc("/default-pet", fetch.ToHandlerFunc(func(_ fetch.Empty) (Pet, error) {
     return Pet{Name: "Teddy"}, nil
 }))
 ```
-If you need to access http request attributes wrap the input with `fetch.Request`:
+If you need to access http request attributes wrap the input with `fetch.Request`. `http.Request` will be embedded to the input.
 ```go
 type Pet struct {
     Name string
 }
-http.HandleFunc("POST /pets/{id}", fetch.ToHandlerFunc(func(in fetch.Request[Pet]) (fetch.Empty, error) {
-    fmt.Println("Pet's id from url:", in.PathValue("id"))
+http.HandleFunc("/pets", fetch.ToHandlerFunc(func(in fetch.Request[Pet]) (*fetch.Empty, error) {
     fmt.Println("Request context:", in.Context())
     fmt.Println("Authorization header:", in.Headers["Authorization"])
 	fmt.Println("Pet:", in.Body)
-    return fetch.Empty{}, nil
+	fmt.Println("Pet's name:", in.Body.Name)
+    return nil, nil
+}))
+```
+If you have go1.22 and above you can access the wildcards as well.
+```go
+http.HandleFunc("GET /pets/{id}", fetch.ToHandlerFunc(func(in fetch.Request[fetch.Empty]) (*fetch.Empty, error) {
+    fmt.Println("id from url:", in.PathValue("id"))
+    return nil, nil
 }))
 ```
 The error format can be customized with the `fetch.SetRespondErrorFormat` global setter.
