+++ go.mod
@@ -1,3 +1,3 @@
 module github.com/glossd/fetch
 
-go 1.22
+go 1.21
